data-pipeline/requirements-python-only.txt
openai>=1.0
faiss-cpu>=1.7
pypdf>=3.9
numpy>=1.24
